+++ cre.vcf2db.R
@@ -1,17 +1,10 @@
 # variant report generator
 # Rscript ~/cre/cre.vcf2.db.R <family> noncoding|default=NULL,coding
-add_placeholder <- function(variants, column_name, placeholder){
-    variants[,column_name] <- with(variants, placeholder)
-    return(variants)
-}
-
-get_variants_from_file <- function (filename){
-    variants <- read.delim(filename, stringsAsFactors = F)
-    return(variants)
-}
+# description of columns: 
+# https://docs.google.com/document/d/1zL4QoINtkUd15a0AK4WzxXoTWp2MRcuQ9l_P9-xSlS4/edit?usp=sharing
 
 # returns Hom / Het / - (for HOM reference)
-genotype2zygocity <- function (genotype_str, ref){
+genotype2zygosity <- function (genotype_str, ref){
     # test
     # genotype_str = "A|A|B"
     # genotype_str = "./." - call not possible
@@ -21,10 +14,9 @@ genotype2zygocity <- function (genotype_str, ref){
     # greedy
     genotype_str <- gsub("|", "/", genotype_str, fixed = T)
     genotype_str <- gsub("./.", "Insufficient_coverage", genotype_str, fixed = T)
-    #genotype_str = gsub(".","NO_CALL",genotype_str,fixed=T)
-      
+    
     if(grepl("Insufficient_coverage", genotype_str)){
-      result <- genotype_str
+        result <- genotype_str
     }else{
         ar <- strsplit(genotype_str, "/", fixed = T)
         len <- length(ar[[1]])
@@ -43,132 +35,116 @@ genotype2zygocity <- function (genotype_str, ref){
 # output : family.ensemble.txt
 create_report <- function(family, samples){
     file <- paste0(family, ".variants.txt")
-    variants <- get_variants_from_file(file)
+    variants <- read_delim(file, delim = "\t", col_types = cols(.default = "c"))
     
     impact_file <- paste0(family, ".variant_impacts.txt")
-    impacts <- get_variants_from_file(impact_file)
+    impacts <- read_delim(impact_file, delim = "\t", col_types = cols(.default = "c"))
     
-    # temporarily due to https://github.com/quinlan-lab/vcf2db/issues/48
-    # fixed in vcf2db
-    #transcripts_genes = read.csv("~/cre/data/genes.transcripts.csv")
-    #variants$Ensembl_gene_id=NULL
-    #variants$Ensembl_transcript_id1=variants$Ensembl_transcript_id
+    variants$Ensembl_gene_id <- NULL
+    genes_transcripts <- read_csv("~/cre/data/genes.transcripts.ens_only.csv") %>% 
+        select(-external_gene_name)
     
-    #for (i in 1:nrow(variants)){
-    #    variant_id = variants[i,"Variant_id"]
-    #    #variant_id="4489"
-    #    variant_impacts = subset(impacts, variant_id == variant_id & gene == variants[i,"Gene"])
-    #    variant_impacts = variant_impacts[order(variant_impacts$transcript),]
-        
-    #    if (nrow(variant_impacts)>0){
-    #        variants[i,"Ensembl_transcript_id1"] = variant_impacts[1,"transcript"]
-    #    }
-    #    ar = strsplit(variants[i,"Ensembl_transcript_id1"],".",fixed=T)
-    #    variants[i,"Ensembl_transcript_id1"] = ar[[1]][1]
-    #}
-    
-    #variants = merge(variants,transcripts_genes,by.x='Ensembl_transcript_id1',by.y="Ensembl_transcript_id",all.x=T)
+    variants <- variants %>% left_join(genes_transcripts, by = c("Ensembl_transcript_id" = "Ensembl_transcript_id"))
     
     #Column1 - Position
-    variants$Position <- with(variants, paste(Chrom, Pos, sep = ':'))
+    variants <- variants %>% mutate(Position = paste(Chrom, Pos, sep = ':'))
     
     #Column2 - UCSC link
     sUCSC1 <- "=HYPERLINK(\"http://genome.ucsc.edu/cgi-bin/hgTracks?db=hg19&hgt.out3=10x&position="
     sUCSC2 <- "\",\"UCSC_link\")"
-    variants$UCSC_Link <- with(variants, paste(sUCSC1, Position, sUCSC2, sep = ''))
+    
+    variants <- variants %>% mutate(UCSC_Link = paste(sUCSC1, Position, sUCSC2, sep = ''))
 
     # Column3 = GNOMAD_Link
-    variants$GNOMAD_POS <- with(variants, paste(Chrom,Pos,Ref,Alt, sep='-'))
     sGNOMAD1 <- "=HYPERLINK(\"http://gnomad.broadinstitute.org/variant/"
     sGNOMAD2 <- "\",\"GNOMAD_link\")"
-    variants$GNOMAD_Link <- with(variants, paste(sGNOMAD1, GNOMAD_POS, sGNOMAD2, sep = ''))
+    variants <- variants %>%
+        mutate(GNOMAD_POS = paste(Chrom, Pos, Ref, Alt, sep='-')) %>% 
+        mutate(GNOMAD_Link = paste(sGNOMAD1, GNOMAD_POS, sGNOMAD2, sep = ''))
 
     # Columns 4,5: Ref,Alt
 
     # Column6 - Gene
     variants$Gene[variants$Gene == ""] <- NA
     
-    # Column 6 - Zygosity, column 8 - Burden
-    # use new loader vcf2db.py - with flag  to load plain text
-    # for genotype and depth - Noah
-    # otherwise have to decode BLOB 
-    # snappy decompression
-    # https://github.com/arq5x/gemini/issues/700
-    # https://github.com/lulyon/R-snappy
+    # Column 6 - Zygosity, 
+    # column 8 - Burden
     for(sample in samples){
-        #DEBUG: gene = IL20RA
-        #sample=samples[1]
+        #DEBUG: 
+        #gene = IL20RA
+        #sample <- samples[1]
         
-        zygocity_column_name <- paste0("Zygosity.", sample)
-        #t = lapply(variants[,paste0("gts.",sample),"Ref"],genotype2zygocity)
-        #t = lapply(variants[,paste0("gts.",sample),"Ref"],genotype2zygocity)
-        t <- unlist(mapply(genotype2zygocity, variants[,paste0("gts.",sample)], 
-                           variants[,"Ref"]))
-        variants[,zygocity_column_name] <- unlist(t)
-    
+        zygosity_column_name <- paste0("Zygosity.", sample)
+        genotype_column_name <- paste0("gts.", sample)
+        v_genotype2zygosity <- Vectorize(genotype2zygosity)
+        
+        variants <- variants %>% mutate(!!zygosity_column_name := v_genotype2zygosity(pull(variants,genotype_column_name), Ref))
+        
         burden_column_name <- paste0("Burden.", sample)
         # calculating Burden using gene rather then Ensembl_gene_id - request from Matt
-        t <- subset(variants, 
-                    get(zygocity_column_name) == 'Hom' | get(zygocity_column_name) == 'Het',
-                    select = c("Gene", zygocity_column_name))
-        # counts from plyr
-        df_burden <- count(t, "Gene")    
-        colnames(df_burden)[2] <- burden_column_name
-        variants <- merge(variants, df_burden, all.x = T)
-        variants[,burden_column_name][is.na(variants[,burden_column_name])] <- 0
-        variants[,burden_column_name][is.na(variants$Gene)] <-0
+        burden <- variants %>% 
+            filter(pull(variants, zygosity_column_name) == 'Hom' | pull(variants, zygosity_column_name) == 'Het') %>% 
+            dplyr::select(Gene) %>% 
+            group_by(Gene) %>% summarise(!!burden_column_name := n()) %>% filter(!is.na(Gene))
+        
+        variants <- variants %>% left_join(burden, by = c("Gene" = "Gene"))
+        
+        variants <- variants %>% mutate(!!burden_column_name := replace_na(pull(variants, burden_column_name), 0))
+        variants$Gene <- variants$Gene %>% replace_na("0")
     }
     
     # Column9 = gts
     
     # Column10 = Variation
     
-    # Column11 =  Info
-    variants <- add_placeholder(variants, "Info", "Info")
-    
+    # Column11 =  Info - clean up!
+    variants <- add_column(variants, Info = rep("Info", nrow(variants)))
     for (i in 1:nrow(variants)){
-        #debug: i=1  
-        v_id <- variants[i,"Variant_id"]
-        gene <- variants[i,"Gene"]
-        #for WES reports we need only coding impacts in the info field, for WGS we need all
-        if (coding){
-            gene_impacts <- subset(impacts, variant_id == v_id & is_coding == 1,
-                                   select = c("exon", "hgvsc", "hgvsp"))
-        }else{
-            gene_impacts <- subset(impacts, variant_id == v_id, 
-                                   select = c("exon", "hgvsc", "hgvsp"))
-        }
+        # debug: 
+        # i=3
+        v_id <- pull(variants, Variant_id)[i]
+        gene <- variants[i, "Gene"]
+        # for WES reports we need only coding impacts in the info field, for WGS we need all
+        # but exon, hgvcs, hgvsp are only for coding?
+        coding_impacts <- impacts %>%
+            filter(variant_id == v_id) %>% 
+            select(exon, hgvsc, hgvsp, gene) %>% 
+            filter(!is.na(exon) & !is.na(hgvsc) & !is.na(hgvsp))
         
-        gene_impacts$gene <- rep(gene, nrow(gene_impacts))
-        
-        gene_impacts$exon[gene_impacts$exon==''] <- 'NA'
-        
-        gene_impacts <- gene_impacts[c("gene", "exon", "hgvsc", "hgvsp")]
-        
-        if (nrow(gene_impacts) > 0){
-            v_impacts <- paste0(gene_impacts$gene, ":exon", gene_impacts$exon,
-                                ":", gene_impacts$hgvsc, ":", gene_impacts$hgvsp)
+        if (nrow(coding_impacts) > 0){
+            v_impacts <- paste0(coding_impacts$gene, ":exon", coding_impacts$exon,
+                                ":", coding_impacts$hgvsc, ":", coding_impacts$hgvsp)
             s_impacts <- paste(v_impacts, collapse = ",")
+        }else{
+            highmed_impacts <- impacts %>% 
+                filter(variant_id == v_id, impact_severity %in% c("HIGH", "MED")) %>% 
+                select(gene, hgvsc, spliceregion)
+            if (nrow(highmed_impacts > 0)){
+                v_impacts <- paste0(highmed_impacts$gene, ":", highmed_impacts$hgvsc, ":", 
+                                highmed_impacts$spliceregion)
+                s_impacts <- paste(v_impacts, collapse = ",")
+            }
+            else
+                s_impacts <- NA
         }
-        else s_impacts <- 'NA'
-      
-        variants[i,"Info"] <- s_impacts
+        variants[i, "Info"] <- s_impacts
     }
     
     # Column12 - Refseq_change
-    variants <- add_placeholder(variants, "Refseq_change", "NA")
+    variants <- add_column(variants, Refseq_change = rep("NA", nrow(variants)))
     
     # Columns 13,14 - Depth, Quality
 
     # Column 15 - Alt_depth - from v.gt_alt_depths
     # when multiple callers used, AD is not set and fixed in merge_reports function
     for(sample in samples){
+        old_name <- paste0("gt_alt_depths.", sample)
         new_name <- paste0("Alt_depths.", sample)
-        setnames(variants, paste0("gt_alt_depths.", sample), new_name)
+        variants <- rename(variants, !!new_name := old_name)
     }
 
-    # Column 16 - Trio_coverage - fixed in merge_reports function
-    variants <- add_placeholder(variants, "Trio_coverage", "")
+    # Column 16 - Trio_coverage - fixed in merge_reports function - !cleanup
+    variants <- add_column(variants, Trio_coverage = "")
     n_sample <- 1
     prefix <- ""
     
@@ -176,54 +152,45 @@ create_report <- function(family, samples){
     variants$gts <- ""
     for(sample in samples){
         column <- paste0("gt_depths.", sample)
-        
         if (n_sample>1) prefix <- "/"
-        
         variants$Trio_coverage <- with(variants, paste0(Trio_coverage, prefix, get(column)))
-      
         column <- paste0("gts.", sample)
-        
         if (n_sample>1) prefix <- ","
-        
         variants$gts <- with(variants,paste0(gts, prefix, get(column)))
-      
         n_sample <- n_sample+1
     }
     
     # Column17 = Ensembl_gene_id
 
     # Column18 = Gene_description
-    gene_descriptions <- read.delim2(paste0(default_tables_path,"/ensembl_w_description.txt"), 
-                                     stringsAsFactors = F)
-    variants <- merge(variants, gene_descriptions, by.x = "Ensembl_gene_id",
-                      by.y = "ensembl_gene_id", all.x = T)
-    
+    gene_descriptions <- read_csv(paste0(default_tables_path, "/ensembl_genes_w_description.csv"))
+    variants <- left_join(variants, gene_descriptions, by = c("Ensembl_gene_id" = "ensembl_gene_id")) %>% 
+        rename(Gene_description = gene_description)
     # Column19 - Omim_gene_description
     omim_file_name <- paste0(default_tables_path,"/omim.txt")
     
     if (file.exists(omim_file_name)){
-	    omim <- read.delim2(omim_file_name, stringsAsFactors=F)
-	    variants <- merge(variants, omim, all.x=T)
-	    variants$Omim_gene_description[is.na(variants$Omim_gene_description)] <- 0
+	    omim <- read_tsv(omim_file_name)
+	    variants <- variants %>% left_join(omim, by = c("Ensembl_gene_id" = "Ensembl_gene_id"))
+	    variants$Omim_gene_description <- variants$Omim_gene_description %>% replace_na(0)
     }
         
     # Column20 - Omim_inheritance 
     omim_inheritance_file_name <- paste0(default_tables_path,"/omim.inheritance.csv")        
     
     if (file.exists(omim_inheritance_file_name)){
-	    omim_inheritance <- read.csv(omim_inheritance_file_name, stringsAsFactors = F)
-	    variants <- merge(variants, omim_inheritance, all.x = T)
+	    omim_inheritance <- read_csv(omim_inheritance_file_name)
+	    variants <- variants %>% left_join(omim_inheritance, by = c("Ensembl_gene_id" = "Ensembl_gene_id"))
     }
 
     # Column 21 = Orphanet
     # previous name - orphanet.deduplicated.txt
-    orphanet_file_name <- paste0(default_tables_path,"/orphanet.txt")
+    orphanet_file_name <- paste0(default_tables_path,"/orphanet.csv")
        
     if (file.exists(orphanet_file_name)){
-	    orphanet <- read.delim(orphanet_file_name, stringsAsFactors = F)  
-	    variants <- merge(variants, orphanet, all.x = T)
-    
-	    variants$Orphanet[is.na(variants$Orphanet)] = 0
+	    orphanet <- read_csv(orphanet_file_name)
+	    variants <- variants %>% left_join(orphanet, by = c("Ensembl_gene_id" = "Ensembl_gene_id"))
+	    variants$Orphanet <- variants$Orphanet %>% replace_na(0)
     }
     
     # Column 22 - Clinvar
@@ -231,21 +198,20 @@ create_report <- function(family, samples){
     # Column 23 - Ensembl_transcript_id
     
     # Column 24 - AA_position
-    # changing separator from / to _ because otherwise excel converts it into date
-    variants[,"AA_position"] <- with(variants, gsub("/", "_", AA_position), fixed = T)
-    
     # Column 25 - Exon
-    variants[,"Exon"] <- with(variants,gsub("/", "_", Exon), fixed = T)
+    # changing separator from / to _ because otherwise excel converts it into date
+    variants <- variants %>% mutate(AA_position = str_replace(AA_position, "/", "_"),
+                                    Exon = str_replace(Exon, "/", "_"))
     
     # Column 26 - Protein_domains
     
     # Column 27, 28 = Frequency_in_C4R, Seen_in_C4R_samples
-    variants <- add_placeholder(variants, "Frequency_in_C4R", "Frequency_in_C4R")
-    variants <- add_placeholder(variants, "Seen_in_C4R_samples", "Seen_in_C4R_samples")
+    variants <- add_column(variants, Frequency_in_C4R = "Frequency_in_C4R")
+    variants <- add_column(variants, Seen_in_C4R_samples = "Seen_in_C4R_samples")
     
     # Columns 29,30,31,32: HGMD
     for(hgmd_field in c("HGMD_id", "HGMD_gene", "HGMD_tag", "HGMD_ref")){
-        variants <- add_placeholder(variants, hgmd_field, "NA")
+        variants <- add_column(variants, !!hgmd_field := "NA")
     }
 
     # Column 33 - rsIds
@@ -258,8 +224,8 @@ create_report <- function(family, samples){
     # Column36 = Gnomad_oe_lof_score
     # Column37 = Gnomad_oe_mis_score
     gnomad_scores_file <- paste0(default_tables_path, "/gnomad_scores.csv")
-    gnomad_scores <- read.csv(gnomad_scores_file, stringsAsFactors = F)
-    variants <- merge(variants, gnomad_scores, all.x = T, all.y = F)
+    gnomad_scores <- read_csv(gnomad_scores_file)
+    variants <- left_join(variants, gnomad_scores)
 
     # Column38 = Gnomad_ac
     # Column39 = Gnomad_hom
@@ -276,7 +242,7 @@ create_report <- function(family, samples){
     # Column44 = cadd
     # Column45 = vest3
     for (i in 1:nrow(variants)){
-        v_vest <- strsplit(variants[i,"Vest3_score"], ",", fixed = T)[[1]]
+        v_vest <- str_split(variants[i,"Vest3_score"], ",")[[1]]
         variants[i, "Vest3_score"] <- max(v_vest)
     }
     
@@ -286,24 +252,26 @@ create_report <- function(family, samples){
     
     # Column47 = Imprinting_status
     # Column48 = Imprinting_expressed_allele
-    imprinting_file_name <- paste0(default_tables_path, "/imprinting.txt")
-    imprinting <- read.delim(imprinting_file_name, stringsAsFactors = F)
-    variants <- merge(variants, imprinting, all.x = T)
+    imprinting_file_name <- paste0(default_tables_path, "/imprinting.csv")
+    imprinting <- read_csv(imprinting_file_name, col_types = cols(col_character(),
+                                                                  col_character(),
+                                                                  col_character()))
+    variants <- left_join(variants, imprinting, by = c("Gene" = "Gene"))
     
     # Column49 - pseudoautosomal
-    pseudoautosomal_file_name <- paste0(default_tables_path, "/pseudoautosomal.txt")
-    pseudoautosomal <- read.delim(pseudoautosomal_file_name, stringsAsFactors = F)
-    variants <- merge(variants, pseudoautosomal, all.x = T)
+    pseudoautosomal_file_name <- paste0(default_tables_path, "/pseudoautosomal.csv")
+    pseudoautosomal <- read_csv(pseudoautosomal_file_name)
+    variants <- left_join(variants, pseudoautosomal, by = c("Ensembl_gene_id" = "Ensembl_gene_id"))
     
     # Column50 - splicing
-    variants <- add_placeholder(variants, "Splicing", "NA")
+    variants <- add_column(variants, Splicing = "NA")
     if ("spliceregion" %in% colnames(impacts))
     {
         for (i in 1:nrow(variants)){
-	          v_id <- variants[i,"Variant_id"]
-	          splicing_impacts <- subset(impacts, variant_id == v_id,
-	                                    select = c("maxentscan_diff","spliceregion"))
-	          splicing_impacts <- subset(splicing_impacts, !is.na(maxentscan_diff))
+              v_id <- variants[i, "Variant_id"]
+	          splicing_impacts <- impacts %>% filter(variant_id == v_id) %>% 
+	                                    select(maxentscan_diff, spliceregion) %>% 
+	                                    filter(!is.na(maxentscan_diff))
 	          splicing_impacts <- unique(splicing_impacts[order(splicing_impacts$maxentscan_diff),])
 	          # capture the absolute difference - very weak site, or very strong site
 	          # negative - strong alt, + weak alt.
@@ -311,21 +279,23 @@ create_report <- function(family, samples){
 	          s_splicing_field <- 0
 	          
 	          if (nrow(splicing_impacts) > 0){
-	              strongest_alt_site <- head(splicing_impacts, n=1)
+	              strongest_alt_site <- head(splicing_impacts, n = 1)
 	              s_splicing_field <- strongest_alt_site$maxentscan_diff
+	              #cat(s_splicing_field)
 	          }
 	          
 	          if (nrow(splicing_impacts) > 1){
 	              weakest_alt_site <- tail(splicing_impacts, n=1)
 	              s_splicing_field <- paste0(s_splicing_field, ";", weakest_alt_site$maxentscan_diff)
+	              #cat(s_splicing_field)
 	          }
 	          
-	          variants[i,"Splicing"] <- s_splicing_field
+	          variants[i, "Splicing"] <- s_splicing_field
         }
     }else print("VEP MaxEntScan annotation is missing")
     
     # Column 51: number of callers
-    variants <- add_placeholder(variants, "Number_of_callers", "Number_of_callers")
+    variants <- add_column(variants, Number_of_callers = "Number_of_callers")
     
     # Column 52: Old multiallelic
     variants$Old_multiallelic[variants$Old_multiallelic == "None"] <- "NA"
@@ -340,32 +310,28 @@ create_report <- function(family, samples){
         variants[,field] <- with(variants, gsub("-1", NA, get(field), fixed = T))  
     }
 
-    print(sort(colnames(variants)))
     select_and_write2(variants, samples, paste0(family, ".create_report"))
 }
 
 # writes in CSV format
 select_and_write2 <- function(variants, samples, prefix)
 {
-    variants <- variants[c(c("Position", "UCSC_Link", "GNOMAD_Link", "Ref", "Alt"),
-                          paste0("Zygosity.", samples),
-                          c("Gene"),
-                          paste0("Burden.", samples),
-                          c("gts", "Variation", "Info", "Refseq_change", "Depth", "Quality"),
-                          paste0("Alt_depths.", samples),
-                          c("Trio_coverage", "Ensembl_gene_id", "Gene_description", "Omim_gene_description", "Omim_inheritance",
-                            "Orphanet", "Clinvar",
-                            "Frequency_in_C4R", "Seen_in_C4R_samples", "HGMD_id", "HGMD_gene", "HGMD_tag", "HGMD_ref",
-                            "Gnomad_af_popmax", "Gnomad_af", "Gnomad_ac", "Gnomad_hom",
-                            "Ensembl_transcript_id", "AA_position", "Exon", "Protein_domains", "rsIDs",
-                            "Gnomad_oe_lof_score", "Gnomad_oe_mis_score", "Exac_pli_score", "Exac_prec_score", "Exac_pnull_score",
-                            "Conserved_in_20_mammals", "Sift_score", "Polyphen_score", "Cadd_score", "Vest3_score", "Revel_score", "Gerp_score",
-                            "Imprinting_status", "Imprinting_expressed_allele", "Pseudoautosomal", "Splicing",
-                            "Number_of_callers", "Old_multiallelic"))]
-  
-    variants <- variants[order(variants$Position),]
+    variants <- variants %>% select(Position, UCSC_Link, GNOMAD_Link, Ref, Alt,
+                                    one_of(paste0("Zygosity.", samples)), Gene, 
+                                    one_of(paste0("Burden.", samples)),
+                                    gts, Variation, Info, Refseq_change, Depth, Quality,
+                                    one_of(paste0("Alt_depths.", samples)),
+                                    Trio_coverage, Ensembl_gene_id, Gene_description, Omim_gene_description, Omim_inheritance,
+                                    Orphanet, Clinvar,
+                                    Frequency_in_C4R, Seen_in_C4R_samples, HGMD_id, HGMD_gene, HGMD_tag, HGMD_ref,
+                                    Gnomad_af_popmax, Gnomad_af, Gnomad_ac, Gnomad_hom,
+                                    Ensembl_transcript_id, AA_position, Exon, Protein_domains, rsIDs,
+                                    Gnomad_oe_lof_score, Gnomad_oe_mis_score, Exac_pli_score, Exac_prec_score, Exac_pnull_score,
+                                    Conserved_in_20_mammals, Sift_score, Polyphen_score, Cadd_score, Vest3_score, Revel_score, Gerp_score,
+                                    Imprinting_status, Imprinting_expressed_allele, Pseudoautosomal, Splicing,
+                                    Number_of_callers, Old_multiallelic)
     
-    write.csv(variants, paste0(prefix,".csv"), row.names = F)
+    write_excel_csv(variants, paste0(prefix,".csv"))
 }
 
 fix_column_name <- function(column_name){
@@ -502,12 +468,13 @@ merge_reports <- function(family, samples){
     platypus_file <- paste0(family, "-platypus-annotated-decomposed.table")
     if(file.exists(platypus_file)){
         platypus <- read.delim(platypus_file, stringsAsFactors = F)
-        platypus$superindex <- with(platypus, paste(paste0(CHROM,":",POS), REF, ALT, sep = '-'))
-        platypus[c("CHROM", "POS", "REF", "ALT")] <- NULL
-        ensemble <- merge(ensemble, platypus, by.x = "superindex", by.y = "superindex", 
+        if (nrow(platypus) > 0){
+            platypus$superindex <- with(platypus, paste(paste0(CHROM,":",POS), REF, ALT, sep = '-'))
+    	    platypus[c("CHROM", "POS", "REF", "ALT")] <- NULL
+    	    ensemble <- merge(ensemble, platypus, by.x = "superindex", by.y = "superindex", 
                           all.x = T, all.y = F)
     
-        for (i in 1:nrow(ensemble)){
+	    for (i in 1:nrow(ensemble)){
           #if(grepl("NA",ensemble[i,"Trio_coverage"])) - wrong, may be 10/10/NA in gatk
           #if (ensemble[i,"Trio_coverage"]=="NA/NA/NA")
             if (str_count(ensemble[i,"Trio_coverage"],"NA") == length(samples)){
@@ -537,6 +504,7 @@ merge_reports <- function(family, samples){
                     n_sample <- n_sample + 1
                 }
           }
+        }
       }
     
       for (sample in samples){
@@ -588,14 +556,13 @@ merge_reports <- function(family, samples){
     select_and_write2(ensemble, samples, paste0(family, ".merge_reports"))
 }
 
-annotate_w_care4rare <- function(family,samples){
-    variants <- read.csv(paste0(family, ".merge_reports.csv"), stringsAsFactors = F)
+annotate_w_care4rare <- function(family, samples){
+    variants <- read_csv(paste0(family, ".merge_reports.csv"), col_types = cols(.default="c"))
   
     variants$superindex <- with(variants, paste(Position, Ref, Alt, sep='-'))
     
     if(exists("seen_in_c4r_counts")){
-        variants <- merge(variants, seen_in_c4r_counts, by.x = "superindex", 
-                          by.y = "Position.Ref.Alt", all.x = T)
+        variants <- left_join(variants, seen_in_c4r_counts, by = c("superindex" = "Position-Ref-Alt"))
         variants$Frequency_in_C4R <- variants$Frequency
         variants$Frequency <- NULL
     }
@@ -603,8 +570,7 @@ annotate_w_care4rare <- function(family,samples){
     variants$Frequency_in_C4R[is.na(variants$Frequency_in_C4R)] <- 0
     
     if(exists("seen_in_c4r_samples")){
-        variants <- merge(variants,seen_in_c4r_samples,by.x = "superindex", 
-                          by.y = "Position.Ref.Alt", all.x = T)
+        variants <- left_join(variants, seen_in_c4r_samples, by = c("superindex" = "Position-Ref-Alt"))
         variants$Seen_in_C4R_samples <- variants$Samples
     }
     
@@ -615,15 +581,13 @@ annotate_w_care4rare <- function(family,samples){
         variants$HGMD_id <- NULL
         variants$HGMD_ref <- NULL
         variants$HGMD_tag <- NULL
-        variants <- merge(variants, hgmd, by.x = "superindex", 
-                          by.y = "superindex", all.x = T, all.y = F)
+        variants <- left_join(variants, hgmd, by = c("superindex" = "superindex"))
         variants$HGMD_gene <- NULL
         
-        hgmd.genes <- as.data.frame(unique(sort(hgmd$HGMD_gene)))
+        hgmd.genes <- tibble(unique(sort(hgmd$HGMD_gene)))
         hgmd.genes <- cbind(hgmd.genes, hgmd.genes)
         colnames(hgmd.genes) <- c("index", "HGMD_gene")
-        variants <- merge(variants, hgmd.genes, by.x = "Gene", by.y = "index",
-                          all.x = T, all.y = F)
+        variants <- left_join(variants, hgmd.genes, by = c("Gene" = "index"))
     }
     
     select_and_write2(variants, samples, paste0(family, ".wes.", Sys.Date()))
@@ -631,73 +595,64 @@ annotate_w_care4rare <- function(family,samples){
 
 load_tables <- function(debug = F){
     print(paste0("Debug:", debug))
-    #debug
-    if (debug == T){
+    if (debug){
         seen_in_c4r_counts.txt <- "seen_in_c4r_counts.txt"
         seen_in_c4r_samples.txt <- "seen_in_c4r_samples.txt"
         hgmd.csv <- "hgmd.csv"
     }else{
-        seen_in_c4r_counts.txt <- paste0(c4r_database_path,"/seen_in_c4r_counts.txt")    
-        seen_in_c4r_samples.txt <- paste0(c4r_database_path,"/seen_in_c4r_samples.txt")
-        hgmd.csv <- paste0(c4r_database_path,"/hgmd.csv")
+        seen_in_c4r_counts.txt <- paste0(c4r_database_path, "/seen_in_c4r_counts.txt")    
+        seen_in_c4r_samples.txt <- paste0(c4r_database_path, "/seen_in_c4r_samples.txt")
+        hgmd.csv <- paste0(c4r_database_path, "/hgmd.csv")
     }
     
     if (file.exists(seen_in_c4r_counts.txt)){
-        seen_in_c4r_counts <<- read.delim(seen_in_c4r_counts.txt, stringsAsFactors=F)
-    }else{
-        print("No C4R counts found")
-    }
+        seen_in_c4r_counts <<- read_delim(seen_in_c4r_counts.txt, delim = "\t")
+    }else print("No C4R counts found")
     
     if (file.exists(seen_in_c4r_samples.txt)){
-        seen_in_c4r_samples <<- read.delim(seen_in_c4r_samples.txt, stringsAsFactors=F)
-    }else{
-        print("No C4R samples found")
-    }
+        seen_in_c4r_samples <<- read_delim(seen_in_c4r_samples.txt, delim = "\t")
+    }else print("No C4R samples found")
     
+    # check hgmd parsing!
     if (file.exists(hgmd.csv)){
-        hgmd <- read.csv(hgmd.csv,stringsAsFactors = F,header = F)
-        colnames(hgmd) <- c("chrom","pos","HGMD_id","ref","alt","HGMD_gene","HGMD_tag","author",
-                           "allname","vol","page","year","pmid")
-        hgmd$superindex <- with(hgmd,paste0(chrom,':',pos,'-',ref,'-',alt))
-        hgmd$HGMD_ref <- with(hgmd,paste(author,allname,vol,page,year,"PMID:",pmid,sep = ' '))
-        hgmd <<- hgmd[,c("superindex","HGMD_id","HGMD_gene","HGMD_tag","HGMD_ref")]
-    }else{
-        print("No HGMD database")
-    }
+        hgmd <<- read_csv(hgmd.csv, 
+                         col_names = c("chrom", "pos", "HGMD_id", "ref", "alt", "HGMD_gene",
+                                       "HGMD_tag", "author", "allname","vol","page", "year", "pmid")) %>% 
+                mutate(superindex = paste0(chrom, ':', pos, '-', ref, '-', alt),
+                       HGMD_ref = paste(author, allname, vol, page, year, "PMID:", pmid, sep = ' ')) %>% 
+                select(c(superindex, HGMD_id, HGMD_gene, HGMD_tag, HGMD_ref))
+    }else print("No HGMD database")
 }
 
 # creates clinical report - more conservative filtering and less columns
-clinical_report <- function(project,samples){
+clinical_report <- function(project, samples){
     report_file_name <- paste0(project,".wes.",Sys.Date(),".csv")
-    full_report <- read.csv(report_file_name, header = T, stringsAsFactors = F)
-    
+    full_report <- read_csv(report_file_name, col_types = cols(.default = "c"))
+    #full_report <- mutate(full_report, max_alt = max(get(paste0("Alt_depths.", samples))))
     full_report$max_alt <- with(full_report, pmax(get(paste0("Alt_depths.", samples))))
     
-    filtered_report <- subset(full_report, 
-               Quality > 1000 & Gnomad_af_popmax < 0.005 & Frequency_in_C4R < 6 & max_alt >=20,
-               select = c("Position", "GNOMAD_Link", "Ref", "Alt", "Gene", paste0("Zygosity.", samples), 
-                          paste0("Burden.",samples),
-                        "Variation", "Info", "Refseq_change", "Omim_gene_description", "Omim_inheritance",
-                        "Orphanet", "Clinvar", "Frequency_in_C4R",
-                        "Gnomad_af_popmax", "Gnomad_af", "Gnomad_ac", "Gnomad_hom",
-                        "Sift_score", "Polyphen_score", "Cadd_score", "Vest3_score", "Revel_score",
-                        "Imprinting_status", "Pseudoautosomal")
-               )
+    # no burden
+    filtered_report <- full_report %>% filter (Quality > 1000 & Gnomad_af_popmax < 0.005 & Frequency_in_C4R < 6 & max_alt >=20) %>% 
+               select(Position, GNOMAD_Link, Ref, Alt, Gene, one_of(paste0("Zygosity.", samples)),
+                          Variation, Info, Refseq_change, Omim_gene_description, 
+                      Omim_inheritance, Orphanet, Clinvar, Frequency_in_C4R, Gnomad_af_popmax, 
+                      Gnomad_af, Gnomad_ac, Gnomad_hom, Sift_score, Polyphen_score, Cadd_score, 
+                      Vest3_score, Revel_score, Imprinting_status, Pseudoautosomal)
     
     # recalculate burden using the filtered report
     for(sample in samples){
         zygosity_column_name <- paste0("Zygosity.", sample)
         burden_column_name <- paste0("Burden.", sample)
-        t <- subset(filtered_report, 
-                    get(zygosity_column_name) == 'Hom' | get(zygosity_column_name) == 'Het',
-                    select = c("Gene", zygosity_column_name))
-        # count is from plyr
-        df_burden <- count(t, "Gene")    
-        colnames(df_burden)[2] <- burden_column_name
-        filtered_report[,burden_column_name] <- NULL
-        filtered_report <- merge(filtered_report, df_burden, all.x = T)
-        filtered_report[,burden_column_name][is.na(filtered_report[, burden_column_name])] <- 0
-        filtered_report[,burden_column_name][is.na(filtered_report$Gene)] <-0
+        # calculating Burden using gene rather then Ensembl_gene_id - request from Matt
+        burden <- filtered_report %>% 
+            filter(pull(filtered_report, zygosity_column_name) == 'Hom' | pull(filtered_report, zygosity_column_name) == 'Het') %>% 
+            dplyr::select(Gene) %>% 
+            group_by(Gene) %>% summarise(!!burden_column_name := n()) %>% filter(!is.na(Gene))
+        
+        filtered_report <- filtered_report %>% left_join(burden, by = c("Gene" = "Gene"))
+        
+        filtered_report <- filtered_report %>% mutate(!!burden_column_name := replace_na(pull(filtered_report, burden_column_name), 0))
+        filtered_report$Gene <- filtered_report$Gene %>% replace_na("0")
     }
     
     #order columns
@@ -709,18 +664,13 @@ clinical_report <- function(project,samples){
       "Sift_score", "Polyphen_score", "Cadd_score", "Vest3_score", "Revel_score",
       "Imprinting_status", "Pseudoautosomal")]
 
-    write.csv(filtered_report, paste0(project, ".wes.clinical.", Sys.Date(), ".csv"), row.names = F)
+    write_excel_csv(filtered_report, paste0(project, ".wes.clinical.", Sys.Date(), ".csv"))
 }
 
-library(stringr)
-library(data.table)
-library(plyr)
+library(tidyverse)
+#library(data.table) #what for?
 default_tables_path <- "~/cre/data"
 c4r_database_path <- "/hpf/largeprojects/ccm_dccforge/dccforge/results/database"
-
-# R substitutes "-" with "." in sample names in columns so fix this in samples.txt
-# sample names starting with letters should be prefixed by X in *.table
-# for correct processing. most of them start with numbers, and R adds X automatically
 
 args <- commandArgs(trailingOnly = T)
 family <- args[1]
@@ -731,11 +681,10 @@ debug <- F
 
 setwd(family)
 
-samples <- unlist(read.table("samples.txt", stringsAsFactors = F))
-samples <- gsub("-", ".", samples)
-    
+samples <- read_lines("samples.txt")
+
 load_tables(debug)
-create_report(family,samples)
+create_report(family, samples)
 merge_reports(family,samples)
 annotate_w_care4rare(family,samples)
 clinical_report(family,samples)