+++ HISTORY.md
@@ -1,3 +1,13 @@
+## 0.0.4 (23 May 2019)
+- fixed a bug in clinvar column
+
+## 0.0.3 (21 May 2019)
+- bug fixes, WES report generated for NA12878 looks ok
+
+## 0.0.1 (15 May 2019)
+- initial release, compatible with bcbio_1.1.5 and gnomad2.1
+
+## previous notes
 - 2018-11-07: new gnomad obseverved/expected scores instead of pLi and exac_missense. cre.gnomad_scores.R
 - 2018-11-02: added back Gerp_score, updated OMIM
 - 2017-09-22: added Info_refseq and Maf_exac to the database report